+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 CHANGELOG
 =========
 
-unreleased
+1.5.0
 ----------
 - Use version 3.20.1 of braintree-web
 - Update browser-detection to v1.6.0